smart-report-writer/requirements.txt
# Web & API
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
starlette==0.40.0
orjson==3.10.7
python-multipart==0.0.12
aiofiles==24.1.0
limits==3.13.0
slowapi==0.1.9

# DB & persistence
SQLAlchemy==2.0.35
psycopg[binary,pool]==3.2.3
alembic==1.13.2

# Storage (S3/MinIO)
boto3==1.35.36
minio==7.2.7

# Workers
celery==5.4.0
redis==5.0.8

# Parsing & ingestion
python-magic==0.4.27
pdfminer.six==20240706
pypdf==5.0.1
python-docx==1.1.2
markdown-it-py==3.0.0
beautifulsoup4==4.12.3
lxml==5.3.0
pandas==2.2.3
openpyxl==3.1.5
pyarrow==17.0.0
pyyaml==6.0.2
pygments==2.18.0
chardet==5.2.0
tika==2.6.0
ebcdic==1.1.1
defusedxml==0.7.1

# Archives & security
libarchive-c==5.1
py7zr==0.21.0
clamd==1.0.2

# OCR/STT (optional; toggle via flags)
pytesseract==0.3.13
pillow==10.4.0
whisper-timestamped==1.15.4
ffmpeg-python==0.2.0

# LLM & retrieval
langchain==0.3.4
langchain-core==0.3.12
langchain-community==0.3.3
faiss-cpu==1.8.0.post1
tiktoken==0.7.0
httpx==0.27.2
openai==1.51.2

# Jinja & rendering
Jinja2==3.1.4
WeasyPrint==62.3

# Metrics & tracing
prometheus-client==0.21.0
opentelemetry-sdk==1.26.0
opentelemetry-exporter-otlp==1.26.0

# Utilities
tenacity==9.0.0
ujson==5.10.0
